[REF] hotel_reservation: split .py files
