+++ hotel_reservation/models/__init__.py
@@ -1,3 +1,7 @@
-# See LICENSE file for full copyright and licensing details.
-
+from . import hotel_folio
 from . import hotel_reservation
+from . import hotel_reservation_line
+from . import hotel_room
+from . import hotel_room_reservation_line
+from . import quick_room_reservation
+from . import room_reservation_summary
