+++ hotel_reservation/models/hotel_reservation.py
@@ -3,119 +3,11 @@
 import time
 from datetime import datetime, timedelta
 
-import pytz
 from dateutil.relativedelta import relativedelta
 
 from odoo import _, api, fields, models
-from odoo.exceptions import UserError, ValidationError
+from odoo.exceptions import ValidationError
 from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT as dt
-
-
-class HotelFolio(models.Model):
-    _inherit = "hotel.folio"
-    _order = "reservation_id desc"
-
-    reservation_id = fields.Many2one(
-        "hotel.reservation", string="Reservation Id"
-    )
-
-    @api.multi
-    def write(self, vals):
-        context = dict(self._context)
-        if not context:
-            context = {}
-        context.update({"from_reservation": True})
-        res = super(HotelFolio, self).write(vals)
-        reservation_line_obj = self.env["hotel.room.reservation.line"]
-        for folio_obj in self:
-            if folio_obj.reservation_id:
-                for reservation in folio_obj.reservation_id:
-                    reservation_obj = reservation_line_obj.search(
-                        [("reservation_id", "=", reservation.id)]
-                    )
-                    if len(reservation_obj) == 1:
-                        for line_id in reservation.reservation_line:
-                            line_id = line_id.reserve
-                            for room_id in line_id:
-                                vals = {
-                                    "room_id": room_id.id,
-                                    "check_in": folio_obj.checkin_date,
-                                    "check_out": folio_obj.checkout_date,
-                                    "state": "assigned",
-                                    "reservation_id": reservation.id,
-                                }
-                                reservation_obj.write(vals)
-        return res
-
-
-class HotelFolioLineExt(models.Model):
-    _inherit = "hotel.folio.line"
-
-    @api.onchange("checkin_date", "checkout_date")
-    def on_change_checkout(self):
-        res = super(HotelFolioLineExt, self).on_change_checkout()
-        hotel_room_obj = self.env["hotel.room"]
-        avail_prod_ids = []
-        hotel_room_ids = hotel_room_obj.search([])
-        for room in hotel_room_ids:
-            assigned = False
-            for line in room.room_reservation_line_ids:
-                if line.status != "cancel":
-                    if (
-                        self.checkin_date
-                        <= line.check_in
-                        <= self.checkout_date
-                    ) or (
-                        self.checkin_date
-                        <= line.check_out
-                        <= self.checkout_date
-                    ):
-                        assigned = True
-                    elif (
-                        line.check_in <= self.checkin_date <= line.check_out
-                    ) or (
-                        line.check_in <= self.checkout_date <= line.check_out
-                    ):
-                        assigned = True
-            if not assigned:
-                avail_prod_ids.append(room.product_id.id)
-        return res
-
-    @api.multi
-    def write(self, vals):
-        """
-        Overrides orm write method.
-        @param self: The object pointer
-        @param vals: dictionary of fields value.
-        Update Hotel Room Reservation line history"""
-        reservation_line_obj = self.env["hotel.room.reservation.line"]
-        room_obj = self.env["hotel.room"]
-        prod_id = vals.get("product_id") or self.product_id.id
-        chkin = vals.get("checkin_date") or self.checkin_date
-        chkout = vals.get("checkout_date") or self.checkout_date
-        is_reserved = self.is_reserved
-        if prod_id and is_reserved:
-            prod_domain = [("product_id", "=", prod_id)]
-            prod_room = room_obj.search(prod_domain, limit=1)
-            if self.product_id and self.checkin_date and self.checkout_date:
-                old_prd_domain = [("product_id", "=", self.product_id.id)]
-                old_prod_room = room_obj.search(old_prd_domain, limit=1)
-                if prod_room and old_prod_room:
-                    # Check for existing room lines.
-                    srch_rmline = [
-                        ("room_id", "=", old_prod_room.id),
-                        ("check_in", "=", self.checkin_date),
-                        ("check_out", "=", self.checkout_date),
-                    ]
-                    rm_lines = reservation_line_obj.search(srch_rmline)
-                    if rm_lines:
-                        rm_line_vals = {
-                            "room_id": prod_room.id,
-                            "check_in": chkin,
-                            "check_out": chkout,
-                        }
-                        rm_lines.write(rm_line_vals)
-        return super(HotelFolioLineExt, self).write(vals)
 
 
 class HotelReservation(models.Model):
@@ -703,581 +595,3 @@ class HotelReservation(models.Model):
                     duration -= 1
         value.update({"duration": duration})
         return value
-
-
-class HotelReservationLine(models.Model):
-    _name = "hotel_reservation.line"
-    _description = "Reservation Line"
-
-    name = fields.Char("Name")
-    line_id = fields.Many2one("hotel.reservation")
-    reserve = fields.Many2many(
-        "hotel.room",
-        "hotel_reservation_line_room_rel",
-        "hotel_reservation_line_id",
-        "room_id",
-        domain="[('isroom','=',True),\
-                               ('categ_id','=',categ_id)]",
-    )
-    categ_id = fields.Many2one("hotel.room.type", "Room Type")
-
-    @api.onchange("categ_id")
-    def on_change_categ(self):
-        """
-        When you change categ_id it check checkin and checkout are
-        filled or not if not then raise warning
-        -----------------------------------------------------------
-        @param self: object pointer
-        """
-        hotel_room_obj = self.env["hotel.room"]
-        hotel_room_ids = hotel_room_obj.search(
-            [("categ_id", "=", self.categ_id.id)]
-        )
-        room_ids = []
-        if not self.line_id.checkin:
-            raise ValidationError(
-                _(
-                    "Before choosing a room,\n You have to \
-                                     select a Check in date or a Check out \
-                                     date in the reservation form."
-                )
-            )
-        for room in hotel_room_ids:
-            assigned = False
-            for line in room.room_reservation_line_ids:
-                if line.status != "cancel":
-                    if (
-                        self.line_id.checkin
-                        <= line.check_in
-                        <= self.line_id.checkout
-                    ) or (
-                        self.line_id.checkin
-                        <= line.check_out
-                        <= self.line_id.checkout
-                    ):
-                        assigned = True
-                    elif (
-                        line.check_in <= self.line_id.checkin <= line.check_out
-                    ) or (
-                        line.check_in
-                        <= self.line_id.checkout
-                        <= line.check_out
-                    ):
-                        assigned = True
-            for rm_line in room.room_line_ids:
-                if rm_line.status != "cancel":
-                    if (
-                        self.line_id.checkin
-                        <= rm_line.check_in
-                        <= self.line_id.checkout
-                    ) or (
-                        self.line_id.checkin
-                        <= rm_line.check_out
-                        <= self.line_id.checkout
-                    ):
-                        assigned = True
-                    elif (
-                        rm_line.check_in
-                        <= self.line_id.checkin
-                        <= rm_line.check_out
-                    ) or (
-                        rm_line.check_in
-                        <= self.line_id.checkout
-                        <= rm_line.check_out
-                    ):
-                        assigned = True
-            if not assigned:
-                room_ids.append(room.id)
-        domain = {"reserve": [("id", "in", room_ids)]}
-        return {"domain": domain}
-
-    @api.multi
-    def unlink(self):
-        """
-        Overrides orm unlink method.
-        @param self: The object pointer
-        @return: True/False.
-        """
-        hotel_room_reserv_line_obj = self.env["hotel.room.reservation.line"]
-        for reserv_rec in self:
-            for rec in reserv_rec.reserve:
-                hres_arg = [
-                    ("room_id", "=", rec.id),
-                    ("reservation_id", "=", reserv_rec.line_id.id),
-                ]
-                myobj = hotel_room_reserv_line_obj.search(hres_arg)
-                if myobj.ids:
-                    rec.write({"isroom": True, "status": "available"})
-                    myobj.unlink()
-        return super(HotelReservationLine, self).unlink()
-
-
-class HotelRoomReservationLine(models.Model):
-    _name = "hotel.room.reservation.line"
-    _description = "Hotel Room Reservation"
-    _rec_name = "room_id"
-
-    room_id = fields.Many2one("hotel.room", string="Room id")
-    check_in = fields.Datetime("Check In Date", required=True)
-    check_out = fields.Datetime("Check Out Date", required=True)
-    state = fields.Selection(
-        [("assigned", "Assigned"), ("unassigned", "Unassigned")], "Room Status"
-    )
-    reservation_id = fields.Many2one("hotel.reservation", string="Reservation")
-    status = fields.Selection(string="state", related="reservation_id.state")
-
-
-class HotelRoom(models.Model):
-    _inherit = "hotel.room"
-    _description = "Hotel Room"
-
-    room_reservation_line_ids = fields.One2many(
-        "hotel.room.reservation.line", "room_id", string="Room Reserve Line"
-    )
-
-    @api.multi
-    def unlink(self):
-        """
-        Overrides orm unlink method.
-        @param self: The object pointer
-        @return: True/False.
-        """
-        for room in self:
-            for reserv_line in room.room_reservation_line_ids:
-                if reserv_line.status == "confirm":
-                    raise ValidationError(
-                        _(
-                            "User is not able to delete the \
-                                            room after the room in %s state \
-                                            in reservation"
-                        )
-                        % (reserv_line.status)
-                    )
-        return super(HotelRoom, self).unlink()
-
-    @api.model
-    def cron_room_line(self):
-        """
-        This method is for scheduler
-        every 1min scheduler will call this method and check Status of
-        room is occupied or available
-        --------------------------------------------------------------
-        @param self: The object pointer
-        @return: update status of hotel room reservation line
-        """
-        reservation_line_obj = self.env["hotel.room.reservation.line"]
-        folio_room_line_obj = self.env["folio.room.line"]
-        now = datetime.now()
-        curr_date = now.strftime(dt)
-        for room in self.search([]):
-            reserv_line_ids = [
-                reservation_line.id
-                for reservation_line in room.room_reservation_line_ids
-            ]
-            reserv_args = [
-                ("id", "in", reserv_line_ids),
-                ("check_in", "<=", curr_date),
-                ("check_out", ">=", curr_date),
-            ]
-            reservation_line_ids = reservation_line_obj.search(reserv_args)
-            rooms_ids = [room_line.id for room_line in room.room_line_ids]
-            rom_args = [
-                ("id", "in", rooms_ids),
-                ("check_in", "<=", curr_date),
-                ("check_out", ">=", curr_date),
-            ]
-            room_line_ids = folio_room_line_obj.search(rom_args)
-            status = {"isroom": True, "color": 5}
-            if reservation_line_ids.ids:
-                status = {"isroom": False, "color": 2}
-            room.write(status)
-            if room_line_ids.ids:
-                status = {"isroom": False, "color": 2}
-            room.write(status)
-            if reservation_line_ids.ids and room_line_ids.ids:
-                raise ValidationError(
-                    _("Please Check Rooms Status for %s.") % (room.name)
-                )
-        return True
-
-
-class RoomReservationSummary(models.Model):
-    _name = "room.reservation.summary"
-    _description = "Room reservation summary"
-
-    name = fields.Char(
-        "Reservation Summary", default="Reservations Summary", invisible=True
-    )
-    date_from = fields.Datetime("Date From")
-    date_to = fields.Datetime("Date To")
-    summary_header = fields.Text("Summary Header")
-    room_summary = fields.Text("Room Summary")
-
-    @api.model
-    def default_get(self, fields):
-        """
-        To get default values for the object.
-        @param self: The object pointer.
-        @param fields: List of fields for which we want default values
-        @return: A dictionary which of fields with values.
-        """
-        if self._context is None:
-            self._context = {}
-        res = super(RoomReservationSummary, self).default_get(fields)
-        # Added default datetime as today and date to as today + 30.
-        from_dt = datetime.today()
-        dt_from = from_dt.strftime(dt)
-        to_dt = from_dt + relativedelta(days=30)
-        dt_to = to_dt.strftime(dt)
-        res.update({"date_from": dt_from, "date_to": dt_to})
-
-        if not self.date_from and self.date_to:
-            date_today = datetime.datetime.today()
-            first_day = datetime.datetime(
-                date_today.year, date_today.month, 1, 0, 0, 0
-            )
-            first_temp_day = first_day + relativedelta(months=1)
-            last_temp_day = first_temp_day - relativedelta(days=1)
-            last_day = datetime.datetime(
-                last_temp_day.year,
-                last_temp_day.month,
-                last_temp_day.day,
-                23,
-                59,
-                59,
-            )
-            date_froms = first_day.strftime(dt)
-            date_ends = last_day.strftime(dt)
-            res.update({"date_from": date_froms, "date_to": date_ends})
-        return res
-
-    @api.multi
-    def room_reservation(self):
-        """
-        @param self: object pointer
-        """
-        mod_obj = self.env["ir.model.data"]
-        if self._context is None:
-            self._context = {}
-        model_data_ids = mod_obj.search(
-            [
-                ("model", "=", "ir.ui.view"),
-                ("name", "=", "view_hotel_reservation_form"),
-            ]
-        )
-        resource_id = model_data_ids.read(fields=["res_id"])[0]["res_id"]
-        return {
-            "name": _("Reconcile Write-Off"),
-            "context": self._context,
-            "view_type": "form",
-            "view_mode": "form",
-            "res_model": "hotel.reservation",
-            "views": [(resource_id, "form")],
-            "type": "ir.actions.act_window",
-            "target": "new",
-        }
-
-    @api.onchange("date_from", "date_to")  # noqa: C901
-    def get_room_summary(self):  # noqa: C901
-        """
-        @param self: object pointer
-         """
-        res = {}
-        all_detail = []
-        room_obj = self.env["hotel.room"]
-        reservation_line_obj = self.env["hotel.room.reservation.line"]
-        folio_room_line_obj = self.env["folio.room.line"]
-        user_obj = self.env["res.users"]
-        date_range_list = []
-        main_header = []
-        summary_header_list = ["Rooms"]
-        if self.date_from and self.date_to:
-            if self.date_from > self.date_to:
-                raise UserError(
-                    _(
-                        "Please Check Time period Date From can't \
-                                   be greater than Date To !"
-                    )
-                )
-            if self._context.get("tz", False):
-                timezone = pytz.timezone(self._context.get("tz", False))
-            else:
-                timezone = pytz.timezone("UTC")
-            d_frm_obj = (
-                datetime.strptime(self.date_from, dt)
-                .replace(tzinfo=pytz.timezone("UTC"))
-                .astimezone(timezone)
-            )
-            d_to_obj = (
-                datetime.strptime(self.date_to, dt)
-                .replace(tzinfo=pytz.timezone("UTC"))
-                .astimezone(timezone)
-            )
-            temp_date = d_frm_obj
-            while temp_date <= d_to_obj:
-                val = ""
-                val = (
-                    str(temp_date.strftime("%a"))
-                    + " "
-                    + str(temp_date.strftime("%b"))
-                    + " "
-                    + str(temp_date.strftime("%d"))
-                )
-                summary_header_list.append(val)
-                date_range_list.append(temp_date.strftime(dt))
-                temp_date = temp_date + timedelta(days=1)
-            all_detail.append(summary_header_list)
-            room_ids = room_obj.search([])
-            all_room_detail = []
-            for room in room_ids:
-                room_detail = {}
-                room_list_stats = []
-                room_detail.update({"name": room.name or ""})
-                if (
-                    not room.room_reservation_line_ids
-                    and not room.room_line_ids
-                ):
-                    for chk_date in date_range_list:
-                        room_list_stats.append(
-                            {
-                                "state": "Free",
-                                "date": chk_date,
-                                "room_id": room.id,
-                            }
-                        )
-                else:
-                    for chk_date in date_range_list:
-                        ch_dt = chk_date[:10] + " 23:59:59"
-                        ttime = datetime.strptime(ch_dt, dt)
-                        c = ttime.replace(tzinfo=timezone).astimezone(
-                            pytz.timezone("UTC")
-                        )
-                        chk_date = c.strftime(dt)
-                        reserline_ids = room.room_reservation_line_ids.ids
-                        reservline_ids = reservation_line_obj.search(
-                            [
-                                ("id", "in", reserline_ids),
-                                ("check_in", "<=", chk_date),
-                                ("check_out", ">=", chk_date),
-                                ("state", "=", "assigned"),
-                            ]
-                        )
-                        if not reservline_ids:
-                            sdt = dt
-                            chk_date = datetime.strptime(chk_date, sdt)
-                            chk_date = datetime.strftime(
-                                chk_date - timedelta(days=1), sdt
-                            )
-                            reservline_ids = reservation_line_obj.search(
-                                [
-                                    ("id", "in", reserline_ids),
-                                    ("check_in", "<=", chk_date),
-                                    ("check_out", ">=", chk_date),
-                                    ("state", "=", "assigned"),
-                                ]
-                            )
-                            for res_room in reservline_ids:
-                                rrci = res_room.check_in
-                                rrco = res_room.check_out
-                                cid = datetime.strptime(rrci, dt)
-                                cod = datetime.strptime(rrco, dt)
-                                dur = cod - cid
-                                if room_list_stats:
-                                    count = 0
-                                    for rlist in room_list_stats:
-                                        cidst = datetime.strftime(cid, dt)
-                                        codst = datetime.strftime(cod, dt)
-                                        rm_id = res_room.room_id.id
-                                        ci = rlist.get("date") >= cidst
-                                        co = rlist.get("date") <= codst
-                                        rm = rlist.get("room_id") == rm_id
-                                        st = rlist.get("state") == "Reserved"
-                                        if ci and co and rm and st:
-                                            count += 1
-                                    if count - dur.days == 0:
-                                        c_id1 = user_obj.browse(self._uid)
-                                        c_id = c_id1.company_id
-                                        con_add = 0
-                                        amin = 0.0
-                                        if c_id:
-                                            con_add = c_id.additional_hours
-                                        # When configured_addition_hours is
-                                        # greater than zero then we calculate
-                                        # additional minutes
-                                        if con_add > 0:
-                                            amin = abs(con_add * 60)
-                                        hr_dur = abs(dur.seconds / 60)
-                                        # When additional minutes is greater
-                                        # than zero then check duration with
-                                        # extra minutes and give the room
-                                        # reservation status is reserved or
-                                        # free
-                                        if amin > 0:
-                                            if hr_dur >= amin:
-                                                reservline_ids = True
-                                            else:
-                                                reservline_ids = False
-                                        else:
-                                            if hr_dur > 0:
-                                                reservline_ids = True
-                                            else:
-                                                reservline_ids = False
-                                    else:
-                                        reservline_ids = False
-                        fol_room_line_ids = room.room_line_ids.ids
-                        chk_state = ["draft", "cancel"]
-                        folio_resrv_ids = folio_room_line_obj.search(
-                            [
-                                ("id", "in", fol_room_line_ids),
-                                ("check_in", "<=", chk_date),
-                                ("check_out", ">=", chk_date),
-                                ("status", "not in", chk_state),
-                            ]
-                        )
-                        if reservline_ids or folio_resrv_ids:
-                            room_list_stats.append(
-                                {
-                                    "state": "Reserved",
-                                    "date": chk_date,
-                                    "room_id": room.id,
-                                    "is_draft": "No",
-                                    "data_model": "",
-                                    "data_id": 0,
-                                }
-                            )
-                        else:
-                            room_list_stats.append(
-                                {
-                                    "state": "Free",
-                                    "date": chk_date,
-                                    "room_id": room.id,
-                                }
-                            )
-
-                room_detail.update({"value": room_list_stats})
-                all_room_detail.append(room_detail)
-            main_header.append({"header": summary_header_list})
-            self.summary_header = str(main_header)
-            self.room_summary = str(all_room_detail)
-        return res
-
-
-class QuickRoomReservation(models.TransientModel):
-    _name = "quick.room.reservation"
-    _description = "Quick Room Reservation"
-
-    partner_id = fields.Many2one(
-        "res.partner", string="Customer", required=True
-    )
-    check_in = fields.Datetime("Check In", required=True)
-    check_out = fields.Datetime("Check Out", required=True)
-    room_id = fields.Many2one("hotel.room", "Room", required=True)
-    warehouse_id = fields.Many2one("stock.warehouse", "Hotel", required=True)
-    pricelist_id = fields.Many2one("product.pricelist", "pricelist")
-    partner_invoice_id = fields.Many2one(
-        "res.partner", "Invoice Address", required=True
-    )
-    partner_order_id = fields.Many2one(
-        "res.partner", "Ordering Contact", required=True
-    )
-    partner_shipping_id = fields.Many2one(
-        "res.partner", "Delivery Address", required=True
-    )
-    adults = fields.Integer("Adults", size=64)
-
-    @api.onchange("check_out", "check_in")
-    def on_change_check_out(self):
-        """
-        When you change checkout or checkin it will check whether
-        Checkout date should be greater than Checkin date
-        and update dummy field
-        -----------------------------------------------------------
-        @param self: object pointer
-        @return: raise warning depending on the validation
-        """
-        if self.check_out and self.check_in:
-            if self.check_out < self.check_in:
-                raise ValidationError(
-                    _(
-                        "Checkout date should be greater \
-                                         than Checkin date."
-                    )
-                )
-
-    @api.onchange("partner_id")
-    def onchange_partner_id_res(self):
-        """
-        When you change partner_id it will update the partner_invoice_id,
-        partner_shipping_id and pricelist_id of the hotel reservation as well
-        ---------------------------------------------------------------------
-        @param self: object pointer
-        """
-        if not self.partner_id:
-            self.partner_invoice_id = False
-            self.partner_shipping_id = False
-            self.partner_order_id = False
-        else:
-            addr = self.partner_id.address_get(
-                ["delivery", "invoice", "contact"]
-            )
-            self.partner_invoice_id = addr["invoice"]
-            self.partner_order_id = addr["contact"]
-            self.partner_shipping_id = addr["delivery"]
-            self.pricelist_id = self.partner_id.property_product_pricelist.id
-
-    @api.model
-    def default_get(self, fields):
-        """
-        To get default values for the object.
-        @param self: The object pointer.
-        @param fields: List of fields for which we want default values
-        @return: A dictionary which of fields with values.
-        """
-        if self._context is None:
-            self._context = {}
-        res = super(QuickRoomReservation, self).default_get(fields)
-        if self._context:
-            keys = self._context.keys()
-            if "date" in keys:
-                res.update({"check_in": self._context["date"]})
-            if "room_id" in keys:
-                roomid = self._context["room_id"]
-                res.update({"room_id": int(roomid)})
-        return res
-
-    @api.multi
-    def room_reserve(self):
-        """
-        This method create a new record for hotel.reservation
-        -----------------------------------------------------
-        @param self: The object pointer
-        @return: new record set for hotel reservation.
-        """
-        hotel_res_obj = self.env["hotel.reservation"]
-        for res in self:
-            rec = hotel_res_obj.create(
-                {
-                    "partner_id": res.partner_id.id,
-                    "partner_invoice_id": res.partner_invoice_id.id,
-                    "partner_order_id": res.partner_order_id.id,
-                    "partner_shipping_id": res.partner_shipping_id.id,
-                    "checkin": res.check_in,
-                    "checkout": res.check_out,
-                    "warehouse_id": res.warehouse_id.id,
-                    "pricelist_id": res.pricelist_id.id,
-                    "adults": res.adults,
-                    "reservation_line": [
-                        (
-                            0,
-                            0,
-                            {
-                                "reserve": [(6, 0, [res.room_id.id])],
-                                "name": (
-                                    res.room_id and res.room_id.name or ""
-                                ),
-                            },
-                        )
-                    ],
-                }
-            )
-        return rec
